Extract empty section scanner
