+++ src/database/section.rs
@@ -29,21 +29,6 @@ impl Section {
         format!("#{}", self.human_title().to_kebab_case())
     }
 
-    /// populates the given issues list if this section has no content
-    pub fn check_empty(&self, path: &Path, issues: &mut Vec<Issue>) {
-        if !self.body.iter().any(|line| !line.text.is_empty()) {
-            issues.push(Issue::EmptySection {
-                location: Location {
-                    file: path.into(),
-                    line: self.line_number,
-                    start: 0,
-                    end: self.title_line.text.len() as u32,
-                },
-                title: self.human_title().into(),
-            });
-        }
-    }
-
     /// populates the given issues list if this section has an empty title
     pub fn check_empty_title(&self, path: &Path, issues: &mut Vec<Issue>) {
         if self.human_title().is_empty() {
@@ -240,80 +225,6 @@ mod tests {
         for (give, want) in tests {
             let section = Section::with_title(give);
             assert_eq!(section.anchor(), want);
-        }
-    }
-    mod check_empty {
-        use crate::{Document, Issue, Location};
-        use indoc::indoc;
-        use std::path::PathBuf;
-
-        #[test]
-        fn empty_section() {
-            let content = indoc! {"
-            # test document
-
-            ### empty section
-            ### next section
-
-            content"};
-            let doc = Document::from_str("test.md", content).unwrap();
-            let mut have = vec![];
-            for section in doc.content_sections {
-                section.check_empty(&PathBuf::from("test.md"), &mut have);
-            }
-            let want = vec![Issue::EmptySection {
-                location: Location {
-                    file: PathBuf::from("test.md"),
-                    line: 2,
-                    start: 0,
-                    end: 17,
-                },
-                title: "empty section".into(),
-            }];
-            pretty::assert_eq!(have, want);
-        }
-
-        #[test]
-        fn blank_line() {
-            let content = indoc! {"
-            # test document
-
-            ### empty section
-
-            ### next section
-
-            content"};
-            let doc = Document::from_str("test.md", content).unwrap();
-            let mut have = vec![];
-            for section in doc.content_sections {
-                section.check_empty(&PathBuf::from("test.md"), &mut have);
-            }
-            let want = vec![Issue::EmptySection {
-                location: Location {
-                    file: PathBuf::from("test.md"),
-                    line: 2,
-                    start: 0,
-                    end: 17,
-                },
-                title: "empty section".into(),
-            }];
-            pretty::assert_eq!(have, want);
-        }
-
-        #[test]
-        fn content() {
-            let content = indoc! {"
-            # test document
-
-            ### section with content
-
-            content"};
-            let doc = Document::from_str("test.md", content).unwrap();
-            let mut have = vec![];
-            for section in doc.content_sections {
-                section.check_empty(&PathBuf::from("test.md"), &mut have);
-            }
-            assert!(have.is_empty());
         }
     }
