+++ src/scan/mod.rs
@@ -2,6 +2,7 @@
 //! and reports problems found as `Issues`.
 
 pub(crate) mod duplicate_sections;
+pub(crate) mod empty_sections;
 pub(crate) mod footnotes;
 pub(crate) mod links;
 pub(crate) mod section_capitalization;
